requirements/testing-win.txt
-r base.txt

Unipath==1.0
coverage==3.7.1
nose==1.3.4
pep8==1.5.7
dataset==0.5.5
django-haystack==2.3.1
elasticsearch==1.2.0
coveralls==0.5
django-debug-toolbar==1.2.2
biopython==1.65
pytz==2014.10
